Speed up Travis CI Builds (#77)
Builds are currently taking ~20 minutes which is excessive for a library of this size. It appears to be spending about 75% of the time downloading and uploading cache. From what I have read, not caching `target/` for libraries without a `Cargo.lock` makes a lot of sense as it grows really quickly and doesn't provide much benefits.

Ironically builds with cache misses (because cache was cleared) were much faster. [This one](https://travis-ci.org/gnosis/ethcontract-rs/jobs/626203145) lasted only ~8.5 minutes.

### Changes

- [x] No longer cache `target/`: this is the biggest change as the target directory grows very quickly and causes caches to grow to around 8GB. This this the major cause of slow down.
- [x] No longer cache `.cargo/registry`: this also grows as dependencies get resolved to different versions and it can be argued that it is better to download from crates.io as then only the required sources are needed. Additionally, there is no point in keeping a cached crates.io index as this repo does not have a `Cargo.lock` and so will download the index on every build anyway.
- [x] Use sccache: This caches compiled rust artifacts, similar to caching `target/` but in a way that is optimized for cloud storage. This allows dependencies to not be rebuilt while keeping the cache significantly smaller than when caching `target/`

### Test Plan

The average build time for the last 10 builds without cache misses at the time of writing was 19m 57s, the plan is to lower this number.

### Commit History

* don't cache target/

* try out sccache

* use stable rustc to build tools