Tests should assert that futures resolve immediately (#124)
Right now, tests use `futures::future::block_on` to wait for futures to resolve. This can be problematic if code causes tests to deadlock (because of a bug for example). Instead tests should use `FutureExt::now_or_never` which returns an `Option<Future::Output>` that is `Some(value)` if the future resolved immediately or `None` if it didn't. This asserts that test futures resolve immediately which is the case for all our tests.

We wrap `now_or_never` in an `FutureImmediateExt::immediate` helper method that additionally panics if the future did not resolve immediately for our tests.

Closes #120.

### Test Plan

CI