[Easy] Run examples with Truffle instead of Docker (#129)
The motivation behind this is that the examples are designed specifically to run with `truffle develop` so that the same truffle installation used for building the example contracts can be used to run a development node.

Previously, to make things simpler, we used Docker in the Travis CI to run all of the examples and configured the Docker-ized Ganache to have the same default parameters as `truffle develop`. This is a bit fragile as the CI will not actually verify that updates to the Truffle package will not break the current examples and example assumptions (such as default network ID and default port).

This PR addresses that and makes the CI run the examples as is documented in the README. Alternatively, the README and examples could be updated to use Docker and Ganache, but that would add another additional dependency on Docker. The advantage of using Truffle is that it is already needed for building the examples.

### Test plan

CI

### Commit History

* Run examples with truffle instead of docker

* use --cwd flag with yarn instead of cd

* hide stdout

* try and keep STDIN open 4eva

* trying something crazy with named pipes

* cleanup travis file

* try with tail dev/null

* added comment to travis file

* update comments for better English