Yarn upgrade (#273)
For some reason, dependable hasn't been finding these updates to yarn dependencies. This PR just does a yarn upgrade to make sure all the NPM packages in the example truffle project are up to date. I didn't think it was worth splitting this out into multiple PRs as these dependencies are just used for compiling the truffle project for the examples, and not in the actual library code.

### Test Plan

Make sure CI still runs (which compiles and runs all the examples).