Run tests with Github actions
Deployments are still handled by travis because we need to move the
secrets to github.
Likewise some examples need secrets to run. We detect this in the script
so that those are only run when available. This is useful for pull
requests from forks because those do not have access to secrets.
Also made a small code change to fix a nightly clippy warnings.