Restructure crate to have virtual workspace at the top level
According to the cargo documentation
https://doc.rust-lang.org/cargo/reference/workspaces.html the top level
Cargo.toml can include both the crate and the workspace or just the
workspace which is called virtual workspace. I find that less confusing
because we publish several packages from the workspace and not just one
primary.
The new structure has been taken from the tokio-rs/tracing repository.
The ethcontract crate has been moved into a folder, the crate folders
have renamed to the exact crate name and the examples have been turned
into a standalone crate.
The example documentation has been moved into it's own crate because
this fits better than including it in the main library.