Allow for BlockNumber based deployment information (#482)
[Geth v1.10.0](https://github.com/ethereum/go-ethereum/releases/tag/v1.10.0) changed their data retention policy to not include transaction inclusion records that are older than one year:

> Geth 1.10.0 contains some changes which remove unnecessary data in the blockchain database. In particular, Geth no longer keeps transaction inclusion info for all transactions, and instead limits the storage of inclusion records to one year. 

This is a problem for the https://github.com/gnosis/dex-services codebase, as it depends on contracts that have been deployed more [than a year ago](https://rinkeby.etherscan.io/tx/0xc5bf4f48747093a79b623c2a2392ccbdb73a3d788e7d1f0f53f640ea18496d90).

The problem in particular happens in `AllEventsBuilder::query_paginated` when we try to map the deployment transaction hash into a block number to decide the earliest point from which to query events.

This PR suggests a fix that changes the deployment information from being purely tx hash based to an enum that can be either a tx hash or a block number. This will allow us to override the deployment hash with the corresponding block number and avoid a `transaction_receipt` RPC call for old contracts.

### Test Plan
Unit tests and manually tested that the override works in `past_events` example.