Add exponential backoff for transaction confirmation polling (#538)
By switching to polling method to confirm transaction, we've introduced a potential issue with confirmation wait times.

On average, a new block appears every seven or so seconds, so we query for current block number with this frequency. This is fine when we need to wait for several blocks. However it we're only waiting for one or two blocks, seven seconds seems excessive. A new block may appear sooner, and we don't want to waste too much time on waiting.

To fix this, we introduce exponential backoff to the polling mechanism. We start with delay of 250ms, and gradually increase it to 7s.

Fix #534