Use Path.join() to create paths instead of Url.path()
We can't assume that the path section of a file:// URL is a valid
operating system path, so instead we join the root and the source using
Rust's std::path API. This still handles absolute paths the same.

This commit resolves issues where we couldn't use ethcontract-derive
on Windows (e.g. gnosis/ethcontract-rs#729)