Add deployed bytecode to contract artifact (#816)
This PR adds the `deployed_bytecode` property to the `Contract` loaded artifact. This allows users to do things like:
- Use `eth_call` with storage overrides for contracts.
- Verify that the on-chain bytecode matches the expected deployed bytecode value.

The main motivation for this change is that we intent on starting to use `eth_call`s for state overrides in Cow Protocol services, and having it be parsed with the contract artifact makes sense.

### Test Plan

Added a new example that tests that the deployed bytecode data in the contract artifact is working.