Call with state overrides (#983)
Using `eth_call` together with state overrides is pretty useful and
widely supported thing. Currently we need to use very ugly workaround in
https://github.com/cowprotocol/services to use state overrides with
`ethcontract-rs`.
In order to make everyone's lives easier it makes the most sense to make
`call_with_state_overrides` a first class citizen in `ethcontract-rs`.

- I added the `StateOverride` struct which defines the format of things
that can be overriden
- I added `call_with_state_overrides` on `MethodCallBuilder` and
`ViewCallBuilder`
- unfortunately does the `web3` crate not have any nice helper functions
for using state overrides so I implemented the call with the underlying
logic that the `web3` crate would have called

### Test Plan
I tested this by pointing the ethcontract dependency in the cowprotocol
repo to a local checkout with this patch and did a manual test to make
sure things work. Given that we are about to completely abandon this
crate I think it's fine to check in a test specific for this.